Move .options() to top-level client only
